Gatsby (#5)
* Add breakpoint-prefixes, start working on landing page

* WIP Move from Jekyll to Gatsby, turn some includes into components

* Move more includes to components

* Move static assets to static directory

These will not always be static, but for now this is clearly the fastes
way to get this working again.

* Make heroes show up, format date in post-teaser

* Bring CSS-variable for gutter back, turn on debug-mode

* Add Helmet, create component for landing-page blocks

* Add code highlighting, load Inter from rsms, extend post-layout

* Add pages and newsletters, some more utility classes

* Show emoji

I copied the mapping from GitHub. As long as I use the names used in
there, I retain the same URLs as before.

This could now even be replaced easily with image-files kept locally, in
the static/-directory. That is for another improvement.

* Replace Breakout with Taper

If I invert how I think about this, it becomes possible using
React-components. They need valid markup, and beginning with a closing
tag is not valid JSX.

* Add breadcrumbs to more pages, move emoji out of newsletter-title

The `showEmoji`-helper is now an Emoji-component. Makes more sense that
way, as that is what the result was anyways.

* Even more breadcrumbs

* Remove old categories-files

* Add projects and courses

* Fix breadcrumb in project-template

* Remove unnecessary semicolons

* Destructure markdownRemark if custom types aren’t possible

I would prefer an `allPost`-query, but the parsed HTML is not available
in `onCreateNode` yet. The workaround with the “type” will have to
remain in place.

* Show videos on course, destructure more values

* Add icon to CTA on landing page

* Style newsletter landing page

Those icons were tough. For now, I am adding height- and width-utilities
as numbers instead of t-shirt sizes until I figure out what sizes I
need.

* WIP Add feed for posts

* Add more destructuring

* Add sitemap

* Adjust latest newsletters to new frontmatter format

* Remove data-directory, fix index on playlists

* Add favicon, run images through ImageOptim

* Adjust links to social media profiles

* Add title to all and description-meta tag to most pages

* Remove icons-styles

* Add rich preview to all pages and templates

* Style landing page a little more

* Slightly adjust sizes in author-block

* Add featured newsletters to newsletter landing page

* Style me-block a little more

* Adjust margins, link from SML-posts to SML-course

* Add meta-data to course page

* Start moving layout-styles out of stylesheets

* Add contact form through StaticKit

* Use Input-component in all forms

* Style list on newsletter landing page, update /about

Taking out references to project-based collaboration for now, as that
has not happened in a while anyways.

* Rewrite some test, add previews to courses

* Style code blocks

* Style tables

Oops, accidentally used `_color` instead of `_background-color`.

* Style lists

* Add some newsletters, remove some courses, update landing page copy

Let’s get this done now.