Use helper-function to build GraphQL-queries
They were 95% identical, with the only difference being the source
instance name we compare by. Might as well remove that duplication in
the spirit of DRY.

The query used for the categories is now separate, as it doesn’t fit the
same pattern. Instead of giving another parameter to the helper
function, we’ll keep it as its own thing for now. Maybe a different
pattern arises later.