Replace helper-styles for body-copy, begin custom remark plugin
This is amazing. I thought I had to mix “old style CSS” and “new style
CSS”, because I had no influence over the markup that remark created. I
would need to add styles to the `li`s and all other tags to be the same
as for the other tags written by hand.

Turns out that’s totally possible with a custom plugin. I’ll have to go
through all existing CSS, but this way I’ll be able to remove 90%+ of
all non-utility-first CSS.