Split allMdx and mdx into custom types
Instead of always querying for EVERYTHING and then filtering by `type`,
this now allows me to directy query for “just newsletters” or “only the
project with this type”.

As a next step, I can add connections between these.

All this, only to be able to add an author to posts.